refactor: infra 패키지 대상 이동
