+++ src/main/java/kitchenpos/common/infra/DefaultPurgomalumClient.java
@@ -1,7 +1,6 @@
 package kitchenpos.common.infra;
 
 import java.net.URI;
-import kitchenpos.common.domain.PurgomalumClient;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
