+++ src/main/java/kitchenpos/order/infra/DefaultKitchenridersClient.java
@@ -2,7 +2,6 @@ package kitchenpos.order.infra;
 
 import java.math.BigDecimal;
 import java.util.UUID;
-import kitchenpos.order.domain.KitchenridersClient;
 import org.springframework.stereotype.Component;
 
 @Component
