+++ src/test/java/kitchenpos/common/infra/FakePurgomalumClient.java
@@ -2,7 +2,6 @@ package kitchenpos.common.infra;
 
 import java.util.Arrays;
 import java.util.List;
-import kitchenpos.common.domain.PurgomalumClient;
 
 public class FakePurgomalumClient implements PurgomalumClient {
     private static final List<String> profanities;
