+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -10,8 +10,8 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
-import kitchenpos.common.domain.PurgomalumClient;
 import kitchenpos.common.infra.FakePurgomalumClient;
+import kitchenpos.common.infra.PurgomalumClient;
 import kitchenpos.menu.domain.model.Menu;
 import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.menu.infra.InMemoryMenuRepository;
